requirements.txt
pymongo==3.11
pandas==1.3.3
smartapi-python
websocket-client
pyotp
requests==2.26.0
dnspython
login
flask

